+++ src/routes/computePlanDetails/hooks/useCancelComputePlan.tsx
@@ -30,11 +30,9 @@ const useCancelComputePlan = (
     const hasCancellableStatus = useMemo(
         (): boolean =>
             computePlan !== null &&
-            [
-                ComputePlanStatus.doing,
-                ComputePlanStatus.todo,
-                ComputePlanStatus.waiting,
-            ].includes(computePlan.status),
+            [ComputePlanStatus.doing, ComputePlanStatus.created].includes(
+                computePlan.status
+            ),
         [computePlan]
     );